--- notes/flutter_web/lib/pages.g.json
@@ -1,34 +1,0 @@
-{
-  "/* notes.g.json */": {
-    "zh-CN": "由github.com/chen56/note生成，note工具会在笔记根目录生成notes.g.json，用于识别您的笔记项目结构, 本文件应被保存在版本控制中，比如git，不应编辑、移动此文件。 ",
-    "en": "Generated by github.com/chen56/note, the note tool will generate a notes.g.json at the root of the note project, which is used to identify the structure of your note item, This file should be saved in version control, such as git, and should not be edited or moved."
-  },
-  "currentMaxNoteId": 1,
-  "notes": {
-    "/notes/research/research_parameterized_example": {
-      "displayName": "参数化范例机制",
-      "order": 100,
-      "codeExpand": false
-    },
-    "/notes": {
-      "displayName": "flutter-note笔记",
-      "order": 100,
-      "codeExpand": false
-    },
-    "/notes/x.note_inside/how_note_work": {
-      "displayName": "Notebook机制",
-      "order": 100,
-      "codeExpand": false
-    },
-    "/notes/x.note_inside/1.welcome": {
-      "displayName": "welcome",
-      "order": 10,
-      "codeExpand": false
-    },
-    "/notes/x.note_inside/note_test_self": {
-      "displayName": "test",
-      "order": 10,
-      "codeExpand": false
-    }
-  }
-}
